requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database Connections
redis==5.0.1
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Vector Database
pinecone-client==3.0.2

# AI/ML
huggingface-hub==0.20.2
transformers==4.37.1
torch==2.1.2
sentence-transformers==2.3.1

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7